Right-size services layer: consolidate 27 service files to 19
Merge closely related services that were always imported together:
- mutationPacing + mutationRetry → mutation.ts
- workflowSummary + workflowReadiness → workflow.ts
- sourceSnapshot + sourceSelection + sourceItemPolicy → sourceSnapshot.ts
- verification + verificationGate + verificationReport → verification.ts

Extract duplicated CLI utilities (GlobalOptions, buildGlobalArgs,
appendOptionalArg, parsePositiveInt) from 8 phase files into utils/cli.ts.

Deduplicate TransferViewModel reset logic (acknowledgeCompletion now
calls resetRunState instead of repeating the same assignments).

All 120 tests pass. No behavioral changes.
