+++ src/phases/inventory.ts
@@ -8,19 +8,7 @@ import { launchBrowserSession } from "../browser/launch.js";
 import { loadWatchLaterInventory } from "../browser/youtube/inventory.js";
 import type { InventoryOptions } from "../browser/youtube/inventory.js";
 import { computeInventoryFingerprint } from "../services/sourceSnapshot.js";
-
-function parsePositiveInt(value: string | undefined, fallback: number): number {
-  if (!value) {
-    return fallback;
-  }
-
-  const parsed = Number(value);
-  if (!Number.isFinite(parsed) || parsed <= 0) {
-    return fallback;
-  }
-
-  return Math.floor(parsed);
-}
+import { parsePositiveInt } from "../utils/cli.js";
 
 export async function runInventory(command: Command): Promise<void> {
   const ctx = createRunContext(command, "inventory");
